Add basic statistic Logging

+++ gushim/program.py
@@ -23,6 +23,11 @@ END_ATTRIBUTE_FILE = 'attributes01.csv'
 YESHUV_MASK_FILE = r'support_data/Yeshuvim2015.csv'
 MIN_POPULATION = 2000
 SAVE_GUSHIM_SHAPEFILE = True  #save the gushim to shapefile
+EXPORT_TO_GEOJSON = False
+EXPORT_TO_TOPOJSON = True  #Save also to TopoJSON
+DELETE_GEOJSON = True
+WORKSPACE_FOLDER = 'workspace'
+EXPORT_FOLDER = 'export'
 
 
 def setup_logging(default_path='logging.yaml', default_level=logging.INFO, env_key='LOG_CFG'):
@@ -71,21 +76,17 @@ def main():
     logger = setup_logging(default_level=logging.DEBUG)
     logger.info("Program started")
 
-    workspace_folder = get_or_create_folder('workspace')
-    export_folder = get_or_create_folder('export')
-    logger.debug('The work folder is: {0}'.format(workspace_folder))
-
+    workspace_folder = get_or_create_folder(WORKSPACE_FOLDER)
+    export_folder = get_or_create_folder(EXPORT_FOLDER)
+    logger.debug('The workspace folder is: {0}'.format(workspace_folder))
 
     # Load Yeshuvim mask data into pandas
     base_folder = os.path.abspath(os.path.dirname(__file__))
     path_to_file = os.path.join(base_folder, YESHUV_MASK_FILE)
-    df_yeshuv = pd.read_csv(path_to_file,
-                         header=0,
-                         encoding='utf-8-sig'  # 'utf-8' ,index_col=['GUSH_NUM']
-                         , index_col=False
-                         )
+    df_yeshuv = pd.read_csv(path_to_file, header=0, encoding='utf-8-sig', index_col=False)
     logger.info('Successfully loaded yeshuv file {0} into pandas'.format(YESHUV_MASK_FILE))
 
+    # Download file from MAPI
     # file_name = mapi.get_gushim(workspace_folder, gushim_url)  #include copy
     # # file_name = r'opentaba_gushim_prj/gushim/workspace/file_download.zip'
     # compress.zip_uncompress(file_name, workspace_folder)
@@ -94,40 +95,28 @@ def main():
     # Load attribute data into geopandas
     csv_att_file = get_mapi_uncompress_file(workspace_folder, END_ATTRIBUTE_FILE)
     path_to_file = os.path.join(workspace_folder, csv_att_file)
-    df_att = pd.read_csv(path_to_file,
-                         # names=['GUSH_NUM','GUSH_SUFFI','SUB_GUSH_I','STATUS','STATUS_TEX','LOCALITY_I','LOCALITY_N','REG_MUN_ID','REG_MUN_NA','COUNTY_ID','COUNTY_NAM','REGION_ID','REGION_NAM','WP','IS_ANALITY','ANALITY_DA','SHAPE_Leng','SHAPE_Area','GUSH_TYPE','GUSHID'],
-                         header=0,
-                         # usecols = ['GUSH_NUM', ]
-                         # skiprows=1,
-                         encoding='utf-8-sig'  # 'utf-8' ,index_col=['GUSH_NUM']
-                         , index_col=False
-                         )
+    df_att = pd.read_csv(path_to_file, header=0, encoding='utf-8-sig', index_col=False)
     logger.info('Successfully loaded attribute file {0} into pandas'.format(csv_att_file))
 
     # Join DataFrames
     merged_mask = pd.merge(left=df_att, right=df_yeshuv, how='inner', left_on='LOCALITY_I', right_on='ID')
 
-    # Load node data into geopandas
+    # Load node data into Geopandas
     csv_node_file = get_mapi_uncompress_file(workspace_folder, END_NODE_FILE)
-    # geojson_file = geo_utils.csv_to_geojson(csv_node_file[0])
-    # PATH = r"C:\Users\sephi\github\opentaba_gushim\opentaba_gushim_prj\gushim\workspace\subgushall-nodes"
     path_to_file = os.path.join(workspace_folder, csv_node_file)
-    df_node = pd.read_csv(path_to_file
-                          , skiprows=1
-                          , names=['shapeid', 'x', 'y']
-                          , encoding='utf-8-sig'  # 'utf-8'
-                          # , nrows=50000
-                          )
+    df_node = pd.read_csv(path_to_file, skiprows=1, names=['shapeid', 'x', 'y'], encoding='utf-8-sig')
+    # geojson_file = geo_utils.csv_to_geojson(csv_node_file[0]) OLD METHOD to generate GeoJSON - Need to Add Projection
     logger.info('Successfully loaded node file {0} into pandas'.format(csv_node_file))
 
     df_node['geometry'] = df_node.apply(lambda x: (x['x'], x['y']), axis=1)
-    logger.info('Successfully created  point object in dataframe ')
+    logger.info('Successfully created  point coordinates in dataframe ')
 
-    df = df_node.groupby('shapeid')['geometry'].apply(lambda x: Polygon(x.tolist())).reset_index()
+    df_point = df_node.groupby('shapeid')['geometry'].apply(lambda x: Polygon(x.tolist())).reset_index()
     # df_node['geometry'] = df_node.groupby('shapeid').apply(lambda x: Polygon(x.tolist([Point(xy) for xy in zip(x.x, x.y)]))).reset_index()
 
-    df_polygon = gp.GeoDataFrame(df, geometry='geometry')
+    df_polygon = gp.GeoDataFrame(df_point, geometry='geometry')
     df_polygon.crs = {'init': 'epsg:2039'}
+    logger.debug('Successfully created  Polygon object in GeoDataframe ')
 
     # Join DataFrames
     merged_inner = pd.merge(left=merged_mask, right=df_polygon, how='inner', left_on='shapeid', right_on='shapeid')
@@ -138,43 +127,62 @@ def main():
     df_polygon_att = gp.GeoDataFrame(merged_inner, geometry='geometry')
     df_polygon_att.crs = {'init': 'epsg:2039'}  #define GeoDataFrame
     logger.debug('Successfully created GeoDataframe ITM')
-    df_polygon_att_wgs = df_polygon_att.to_crs({'init': 'epsg:4326'}).copy() #world = world.to_crs({'init': 'epsg:3395'})
+    df_polygon_att_wgs = df_polygon_att.to_crs({'init': 'epsg:4326'}).copy()
     logger.info('Successfully projected file GeoDataframe to WGS84')
 
-    #Get list of localities
+    # Get list of localities
     localities = df_polygon_att.groupby(['EngName']).size().index.tolist()
 
-    #export each locality as geojson
+    # export each locality as geojson
     for local in localities:
         df_local = df_polygon_att_wgs[(df_polygon_att_wgs.EngName == local) & (df_polygon_att_wgs.Pop2015 > MIN_POPULATION)]
         if not df_local.empty:
             file_name_string = "".join(x for x in local if x.isalnum()).encode('utf-8')  #remove unsafe characters
             try:
                 export_file = os.path.join(base_folder, export_folder, 'local_{0}.geojson'.format(file_name_string))
-                with open(export_file, 'w') as f:
-                    f.write(df_local.to_json())
+                local_geojson = df_local.to_json()
+                if EXPORT_TO_GEOJSON:
+                    with open(export_file, 'w') as f:
+                        f.write(local_geojson)
+                if EXPORT_TO_TOPOJSON:
+                    if not os.path.exists(export_file) or not os.path.isdir(export_file):
+                        with open(export_file, 'w') as f:
+                            f.write(local_geojson)
+                    filename, _ = os.path.splitext(export_file)
+                    topojson_file = filename + '.topojson'
+                    geo_utils.geoson_to_topojson(export_file, topojson_file)
+                if DELETE_GEOJSON:
+                    try:
+                        os.remove(export_file)
+                    except OSError, e:
+                        print ("Error: {} - {}.".format(e.message, e.strerror))
             except:
                 logger.warning('failed to save geojson for: {0}'.format(local))
     logger.debug('Finished to saved files to GeoJSON')
 
-    # TODO Convert to TopoJSON
-
     # TODO convert to function taking dfpg and crs
 
-    if SAVE_GUSHIM_SHAPEFILE == True:
+    if SAVE_GUSHIM_SHAPEFILE:
         timestr = time.strftime("%Y%m%d_%H%M%S")
         export_file = os.path.join(base_folder, export_folder, 'Gushim{0}.shp'.format(timestr))
         df_polygon_att_wgs.to_file(export_file)
         logger.info('Successfully saved shapefile {0}'.format(export_file))
+
+
     logger.info('Successfully  created  Polygon in geopandas')
 
     # geojson_file = r'C:\Users\sephi\github\opentaba_gushim\opentaba_gushim_prj\gushim\workspace\subgushall-nodes\countries.geo.json'
     # geojson_file = r'C:\Users\sephi\github\opentaba_gushim\opentaba_gushim_prj\gushim\workspace\abugosh.geoson'
-    # topojson_file = os.path.join(os.path.dirname(geojson_file), os.path.basename(geojson_file) + '.topojson')
-    # geo_utils.geoson_to_topojson(geojson_file, topojson_file)
-
+    #
     #Generate Topojson (reuse)
-    #Push to GIT
+
+    # Log Statistics
+    logger.info('Number of Localities to be exported is {0}'.format(len(localities)))
+    logger.info('Number of Gushim / Polygons that have been exported {0}'.format(len(df_polygon_att_wgs.shapeid)))
+    logger.debug('Number of Gushim per Localitis is {0}'.format(df_polygon_att_wgs.groupby(['HebName'])['GUSH_NUM'].size()))
+
+    # TODO Push to GIT
+
     logger.info("Program completed")
